requirements.txt
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.4
opencv-python==4.5.1.48
Pillow==8.4.0
six==1.16.0
torch==1.4.0
torchvision==0.5.0
Werkzeug==2.0.2